cleanup part 1

+++ include/kernel/com/sys/proc.h
@@ -27,8 +27,8 @@
 #include <stdint.h>
 
 typedef struct com_proc {
-    uint64_t              pid;
-    uint64_t              parent_pid;
+    int                   pid;
+    int                   parent_pid;
     bool                  exited;
     int                   exit_status;
     arch_mmu_pagetable_t *page_table;
@@ -38,7 +38,7 @@ typedef struct com_proc {
     _Atomic(com_vnode_t *) cwd;
     // TODO: this is a lock... circular dependencies
     int            fd_lock;
-    uintmax_t      next_fd;
+    int            next_fd;
     com_filedesc_t fd[16];
 
     int    pages_lock;
@@ -48,13 +48,13 @@ typedef struct com_proc {
 } com_proc_t;
 
 com_proc_t *com_sys_proc_new(arch_mmu_pagetable_t *page_table,
-                             uintmax_t             parent_pid,
+                             int                   parent_pid,
                              com_vnode_t          *root,
                              com_vnode_t          *cwd);
 void        com_sys_proc_destroy(com_proc_t *proc);
-uintmax_t   com_sys_proc_next_fd(com_proc_t *proc);
-com_file_t *com_sys_proc_get_file(com_proc_t *proc, uintmax_t fd);
-com_proc_t *com_sys_proc_get_by_pid(uintmax_t pid);
+int         com_sys_proc_next_fd(com_proc_t *proc);
+com_file_t *com_sys_proc_get_file(com_proc_t *proc, int fd);
+com_proc_t *com_sys_proc_get_by_pid(int pid);
 void        com_sys_proc_acquire_glock(void);
 void        com_sys_proc_release_glock(void);
 void        com_sys_proc_wait(com_proc_t *proc);
