+++ src/com/sys/syscall/fork.c
@@ -64,7 +64,7 @@ com_syscall_ret_t com_sys_syscall_fork(arch_context_t *ctx,
         goto end;
     }
 
-    for (size_t i = 0; i < proc->next_fd; i++) {
+    for (int i = 0; i < proc->next_fd; i++) {
         if (NULL != proc->fd[i].file) {
             new_proc->fd[i].flags = proc->fd[i].flags;
             COM_FS_FILE_HOLD(proc->fd[i].file);
